pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["dolphin", "torchql"]

[project]
name = "dolphin"
version = "0.0.1"
description = "A package for writing end-to-end differentiable neurosymbolic programs in Python."
dependencies = [
   "torch", "torchvision", "torchaudio", "deprecation", "tqdm", "numpy-indexed", "wandb", "transformers", "einops", "scipy", "av"
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/aadityanaik/Dolphin"
Issues = "https://github.com/aadityanaik/Dolphin/issues"
